requirements.txt
streamlit>=1.20
opencv-python
pillow
pytesseract
transformers
torch
sentencepiece
python-multipart
numpy
